Update detection code and add landmark predictor
